centralize kb querying
