+++ agent/src/temp/reply-twitter.ts
@@ -1,12 +1,11 @@
 import * as readline from 'node:readline/promises';
-import { getReasonBillContext } from '../twitter/execute-interaction';
 import { CoreMessage, generateText } from 'ai';
 import { REJECTION_REASON, SEED, TEMPERATURE } from '../const';
 import { openai } from '@ai-sdk/openai';
 import { PROMPTS, QUESTION_EXTRACTOR_SYSTEM_PROMPT } from '../twitter/prompts';
 import { generateReply, getTweetContext } from '../twitter/execute';
 import { logger } from '../logger';
-import { getDocumentContext } from '../twitter/helpers';
+import { getKbContext } from '../twitter/knowledge-base';
 
 const log = logger.child({ module: 'cli-reply-twitter' });
 
@@ -61,37 +60,27 @@ async function main() {
     }
     const messages: Array<CoreMessage> = [];
     const fullThread = [...tweetThread, tweetWeRespondingTo!];
-    const documents = await getDocumentContext(
+    const kb = await getKbContext(
       {
         messages: fullThread,
         text: extractedQuestion,
       },
       log,
-    ).catch(_ => {
-      return null;
-    });
+    );
 
-    if (documents) {
+    if (kb?.documents) {
       messages.push({
         role: 'user',
-        content: `Documents Context: ${documents}\n\n`,
+        content: `Documents Context: ${kb.documents}\n\n`,
       });
     }
 
-    const bill = await getReasonBillContext(
-      {
-        messages: fullThread,
-      },
-      log,
-    ).catch(_ => {
-      return null;
-    });
-    const summary = bill ? `${bill.title}: \n\n${bill.content}` : '';
-    if (bill) {
+    const summary = kb?.bill ? `${kb.bill.title}: \n\n${kb.bill.content}` : '';
+    if (kb?.bill) {
       log.info(
         {
-          billId: bill.id,
-          billTitle: bill.title,
+          billId: kb.bill.id,
+          billTitle: kb.bill.title,
         },
         'found bill',
       );
@@ -100,7 +89,7 @@ async function main() {
     if (summary) {
       messages.push({
         role: 'user',
-        content: `Context from database: ${summary}\n\n`,
+        content: `Bills Context: ${summary}\n\n`,
       });
     }
